+++ SudokuBoard.cpp
@@ -1,3 +1,4 @@
+#include "utilityFunctions.h"
 #include "SudokuBoard.h"
 #include "SudokuSolver.h"
 #include <iostream>
@@ -11,27 +12,39 @@ using std::list;
 using std::advance;
 #include <algorithm>
 using std::pair;
+#include <string>
+using std::string;
 
 
 SudokuBoard::GridNumbers SudokuBoard::gridNumbers =  {{
-    {10, 10, 10, 10, 10, 10, 10, 10, 10},
-    {10, 0, 0, 0, 1, 1, 1, 2, 2, 2},
-    {10, 0, 0, 0, 1, 1, 1, 2, 2, 2},
-    {10, 0, 0, 0, 1, 1, 1, 2, 2, 2},
-    {10, 3, 3, 3, 4, 4, 4, 5, 5, 5},
-    {10, 3, 3, 3, 4, 4, 4, 5, 5, 5},
-    {10, 3, 3, 3, 4, 4, 4, 5, 5, 5},
-    {10, 6, 6, 6, 7, 7, 7, 8, 8, 8},
-    {10, 6, 6, 6, 7, 7, 7, 8, 8, 8},
-    {10, 6, 6, 6, 7, 7, 7, 8, 8, 8}
+    {10, 10, 10, 10, 10, 10, 10, 10, 10, 10},
+    {10, 0,  0,  0,  1,  1,  1,  2,  2,  2},
+    {10, 0,  0,  0,  1,  1,  1,  2,  2,  2},
+    {10, 0,  0,  0,  1,  1,  1,  2,  2,  2},
+    {10, 3,  3,  3,  4,  4,  4,  5,  5,  5},
+    {10, 3,  3,  3,  4,  4,  4,  5,  5,  5},
+    {10, 3,  3,  3,  4,  4,  4,  5,  5,  5},
+    {10, 6,  6,  6,  7,  7,  7,  8,  8,  8},
+    {10, 6,  6,  6,  7,  7,  7,  8,  8,  8},
+    {10, 6,  6,  6,  7,  7,  7,  8,  8,  8}
 }};
+
+SudokuBoard::SudokuBoard() {
+    for(size_t i = 0; i < boardSize + 1; i++) {
+        for(size_t j = 0; j < boardSize + 1; j++) {
+            horizontal[i][j] = 0;
+            vertical[i][j] = 0;
+            grids[i][j] = 0;
+        }
+    }
+}
 
 SudokuBoard::SudokuBoard(ifstream &input) {
     for(size_t i = 0; i < boardSize + 1; i++) {
         for(size_t j = 0; j < boardSize + 1; j++) {
-            horizontal[i][j] = false;
-            vertical[i][j] = false;
-            grids[i][j] = false;
+            horizontal[i][j] = 0;
+            vertical[i][j] = 0;
+            grids[i][j] = 0;
         }
     }
 
@@ -47,9 +60,9 @@ SudokuBoard::SudokuBoard(ifstream &input) {
                 board[i][j].empty();
             } else {
                 board[i][j] = value - '0';
-                vertical[j][board[i][j].getValue()] = true;
-                horizontal[i][board[i][j].getValue()] = true;
-                grids[getGrid(i, j)][board[i][j].getValue()] = true;
+                vertical[j][board[i][j].getValue()]++;
+                horizontal[i][board[i][j].getValue()]++;
+                grids[getGrid(i, j)][board[i][j].getValue()]++;
                 numSolved++;
             }
         }
@@ -60,9 +73,9 @@ void SudokuBoard::clearBoard() {
     for(size_t i = 0; i < boardSize; i++) {
         for(size_t j = 0; j < boardSize; j++) {
             board[i][j].empty();
-            horizontal[i + 1][j + 1] = false;
-            vertical[i + 1][j + 1] = false;
-            grids[i + 1][j + 1] = false;
+            horizontal[i + 1][j + 1] = 0;
+            vertical[i + 1][j + 1] = 0;
+            grids[i + 1][j + 1] = 0;
         }
     }
 
@@ -111,9 +124,9 @@ void SudokuBoard::generate(Difficulty difficulty) {
             advance(it, (rand() % valuesRemaining.size()));
 
             board[row][col] = *it;
-            horizontal[row][*it] = true;
-            vertical[col][*it] = true;
-            grids[i][*it] = true;
+            horizontal[row][*it]++;
+            vertical[col][*it]++;
+            grids[i][*it]++;
             numSolved++;
 
             valuesRemaining.erase(it);
@@ -130,18 +143,18 @@ void SudokuBoard::generate(Difficulty difficulty) {
     size_t numToRemove = 0;
     switch(difficulty) {
         case Difficulty::EASY:
-            numToRemove = 40;
+            numToRemove = 35;
         break;
         case Difficulty::MEDIUM:
-            numToRemove = 50;
+            numToRemove = 45;
         break;
 
         case Difficulty::HARD:
-            numToRemove = 60;
+            numToRemove = 50;
         break;
 
         case Difficulty::IMPOSSIBLE:
-            numToRemove = 70;
+            numToRemove = 60;
         break;
 
     }
@@ -169,9 +182,9 @@ void SudokuBoard::generate(Difficulty difficulty) {
 
 void SudokuBoard::tryNum(size_t row, size_t col, size_t val) {
     board[row][col].tryNum(val);
-    horizontal[row][val] = true;
-    vertical[col][val] = true;
-    grids[getGrid(row, col)][val] = true;
+    horizontal[row][val]++;
+    vertical[col][val]++;
+    grids[getGrid(row, col)][val]++;
     numSolved++;
 }
 
@@ -180,81 +193,117 @@ void SudokuBoard::tryIncorrect(size_t row, size_t col, size_t val) {
 
     if(horizontal[row][val]) {
         for(size_t i = 1; i < boardSize + 1; i++) {
-            if(board[row][i] == val) {
+            if(board[row][i] == val && i != col) {
                 board[row][i].breakSquare();
+                numSolved--;
             }
         }
     }
 
     if(vertical[col][val]) {
         for(size_t i = 1; i < boardSize + 1; i++) {
-            if(board[i][col] == val) {
+            if(board[i][col] == val && i != row) {
                 board[i][col].breakSquare();
+                numSolved--;
             }
         }
     }
 
     if(grids[getGrid(row, col)][val]) {
+        size_t minRow = getMinCoord(row);
+        size_t minCol = getMinCoord(col);
+
+        for(size_t i = minRow; i < minRow + 3; i++) {
+            for(size_t j = minCol; j < minCol + 3; j++) {
+                if(board[i][j] == val && (i != row || j != col)) {
+                    board[i][j].breakSquare();
+                    numSolved--;
+                }
+            }
+        }
+    }
+
+    horizontal[row][val]++;
+    vertical[col][val]++;
+    grids[getGrid(row, col)][val]++;
+}
+
+void SudokuBoard::empty(size_t row, size_t col) {
+    if(!board[row][col].occupied()) {
+        return;
+    }
+
+
+    if(board[row][col].incorrect()) {
+        size_t value = board[row][col].getValue();
+
+        //Fix horizontal
+        for(size_t i = 1; i < boardSize + 1; i++) {
+            if(board[row][i] == value && i != col) {
+                board[row][i].fix();
+                numSolved++;
+            }
+        }
+
+        //Fix vertical
+        for(size_t i = 1; i < boardSize + 1; i++) {
+            if(board[i][col] == value && i != row) {
+                board[i][col].fix();
+                numSolved++;
+            }
+        }
+
+
+        //Fix grid
         size_t minRow, minCol;
 
         switch(row) {
-            case 1:
-            case 2:
-            case 3:
+            case 1: case 2: case 3:
             minRow = 1;
             break;
-            case 4:
-            case 5:
-            case 6:
+            case 4: case 5: case 6:
             minRow = 4;
             break;
-            case 7:
-            case 8:
-            case 9:
+            case 7: case 8: case 9:
             minRow = 7;
             break;
         }
 
-        switch(row) {
-            case 1:
-            case 2:
-            case 3:
+        switch(col) {
+            case 1: case 2: case 3:
             minCol = 1;
             break;
-            case 4:
-            case 5:
-            case 6:
+            case 4: case 5: case 6:
             minCol = 4;
             break;
-            case 7:
-            case 8:
-            case 9:
+            case 7: case 8: case 9:
             minCol = 7;
             break;
         }
 
         for(size_t i = minRow; i < minRow + 3; i++) {
             for(size_t j = minCol; j < minCol + 3; j++) {
-                if(board[i][j] == val) {
-                    board[i][j].breakSquare();
+                if(board[i][j] == value && (i != row || j != col)) {
+                    board[i][j].fix();
+                    numSolved++;
                 }
             }
         }
 
+        horizontal[row][board[row][col].getValue()]--;
+        vertical[col][board[row][col].getValue()]--;
+        grids[getGrid(row, col)][board[row][col].getValue()]--;
+
+        board[row][col].empty();
+    } else {
+        horizontal[row][board[row][col].getValue()]--;
+        vertical[col][board[row][col].getValue()]--;
+        grids[getGrid(row, col)][board[row][col].getValue()]--;
+
+        board[row][col].empty();
+        numSolved--;
     }
 
-    horizontal[row][val] = true;
-    vertical[col][val] = true;
-    grids[getGrid(row, col)][val] = true;
-}
-
-void SudokuBoard::empty(size_t row, size_t col) {
-    horizontal[row][board[row][col].getValue()] = false;
-    vertical[col][board[row][col].getValue()] = false;
-    grids[getGrid(row, col)][board[row][col].getValue()] = false;
-
-    board[row][col].empty();
-    numSolved--;
 }
 
 void SudokuBoard::print() {
@@ -279,6 +328,7 @@ void SudokuBoard::print() {
 }
 
 void SudokuBoard::print(size_t row, size_t col) {
+    size_t gridNum = getGrid(row, col);
     for(size_t i = 1; i < boardSize + 1; i++) {
         if((i - 1) % 3 == 0) {
             if(i != 1) {
@@ -291,18 +341,59 @@ void SudokuBoard::print(size_t row, size_t col) {
         }
             for(size_t j = 1; j < boardSize + 1; j++) {
 
+                string color;
                 if(i == row && j == col) {
-                    cout << ((board[i][j].incorrect()) ? "\033[47;31m" : "\033[47m") << board[i][j].getValue() << DEFAULT
+
+                    if(board[i][j].incorrect()) {
+                        color = "\033[47;31m";
+                    } else if(board[i][j].occupied()) {
+                        color = "\033[47;30m";
+                    } else {
+                        color = "\033[47m";
+                    }
+                    cout << color << board[i][j].getValue() << DEFAULT
                     << (((j) % 3 == 0) ? "  #  " : "  ");
                 } else {
-                    cout << board[i][j]
+                    if(row == i || col == j || getGrid(i, j) == gridNum) {
+                        color = "\033[48;5;236m";
+                    }
+
+                    if(board[row][col] == board[i][j].getValue() && board[i][j].getValue()) {
+                        color = "\033[48;2;125;75;10m";
+                    }
+
+                    cout << color <<  board[i][j]
                     << (((j) % 3 == 0) ? "  #  " : "  ");
+                    color = "";
                 }
             }
 
     }
-    cout << "\n#           #           #           #\n"
-         <<   "#####################################\n";
+
+    if(!solved()) {
+        cout << "\n#           #           #           #"
+            << "\n#####################################"
+            << "\n            O P T I O N S            "
+            << "\nArrow Keys   Move around board"
+            << "\nEnter        Focus on selected square"
+            << "\nBackspace    Delete selected square"
+            << "\nQ, q         Quit game";
+    }
+
+    cout << endl;
+}
+
+void SudokuBoard::getSubgrid(const size_t row, const size_t col, BoardSquare *(&grid)[9]) {
+
+    size_t minRow = getMinCoord(row);
+    size_t minCol = getMinCoord(col);
+
+    size_t counter = 0;
+    for(size_t i = minRow; i < minRow + 3; i++) {
+        for(size_t j = minCol; j < minCol + 3; j++) {
+            grid[counter++] = &(board[i][j]);
+        }
+    }
 }
 
 void SudokuBoard::print(ofstream &of) {